+++ ScintSuite/_GUIs/_VideoPlayer.py
@@ -6,7 +6,7 @@ import matplotlib.backends.backend_tkagg as tkagg
 import matplotlib.pyplot as plt
 import ScintSuite._Plotting as ssplt
 import ScintSuite.SimulationCodes.SINPA as sssinpa
-import ScintSuite._Mapping as ssmap
+import ScintSuite._StrikeMap as ssmap
 from matplotlib.figure import Figure
 from tkinter import ttk
 
@@ -15,7 +15,8 @@ class ApplicationShowVid:
     """Class to show the camera frames"""
 
     def __init__(self, master, data, remap_dat, GeomID='AUG02',
-                 calibration=None, scintillator=None, shot: int = None):
+                 calibration=None, scintillator=None, shot: int = None,
+                 mask=None):
         """
         Create the window with the sliders
 
@@ -24,6 +25,7 @@ class ApplicationShowVid:
         :param  remap_dat: the dictionary of remapped data
         :param  GeomID: Geometry id of the detector (to load smaps if needed)
         :param  calibration: Calibration parameters
+        :param  mask: to plot a small coloured mask on top of the image
         """
         # --- List of supported colormaps
         self.cmaps = {
@@ -41,6 +43,7 @@ class ApplicationShowVid:
         self.GeomID = GeomID
         self.CameraCalibration=calibration
         self.scintillator = scintillator
+        self.mask = mask
         t = data['t'].values
         # --- Create a tk container
         frame = tk.Frame(master)
@@ -142,6 +145,20 @@ class ApplicationShowVid:
                                       command=self.scint_Button_change,
                                       takefocus=0, state=state)
         self.scint_button.grid(row=3, column=4)
+        # --- Button for the mask:
+        # If there is not scintillator data, deactivate the button
+        if self.mask is None:
+            state = tk.DISABLED
+        else:
+            state = tk.NORMAL
+        # Initialise the variable of the button
+        self.checkVar3 = tk.BooleanVar()
+        self.checkVar3.set(False)
+        # Create the button
+        self.mask_button = tk.Button(master, text="Draw mask",
+                                      command=self.mask_Button_change,
+                                      takefocus=0, state=state)
+        self.mask_button.grid(row=3, column=5)
         # Draw and show
         self.canvas.draw()
         self.canvas.get_tk_widget().grid(row=0, column=0, columnspan=5,
@@ -151,7 +168,7 @@ class ApplicationShowVid:
         # allows to the canvas to resize when I resize the window
         # --- Quit button
         self.qButton = tk.Button(master, text="Quit", command=master.quit)
-        self.qButton.grid(row=3, column=5)
+        self.qButton.grid(row=5, column=5)
         frame.grid()
 
     def change_cmap(self, cmap_cname):
@@ -182,7 +199,8 @@ class ApplicationShowVid:
             smap_folder = self.remap_dat['frames'].attrs['smap_folder']
             full_name_smap = os.path.join(smap_folder, name__smap)
             # Load the map:
-            smap = ssmap.StrikeMap(0, full_name_smap)
+            # @ToDO: Add here the INPA map if needed
+            smap = ssmap.Fsmap(full_name_smap)
             # Calculate pixel coordinates
             smap.calculate_pixel_coordinates(self.CameraCalibration)
             # Plot the map
@@ -194,6 +212,7 @@ class ApplicationShowVid:
             # Plot the scintillator:
         if self.checkVar2.get():
             self.scintillator.plot_pix(ax=self.canvas.figure.axes[0])
+            # Plot the mask:
         self.canvas.draw()
 
     def set_scale(self):
@@ -221,4 +240,24 @@ class ApplicationShowVid:
         # Now update the value
         self.checkVar2.set(not self.checkVar2.get())
         print('Draw scintillator :', self.checkVar2.get())
+        self.canvas.draw()
+
+    def mask_Button_change(self):
+        """Decide to plot or not the mask. Careful. Press once, move to next
+        frame for it to do something and then deactivate so it doesn't add layers."""
+        
+        # If it was true and we push the button, the smap should be deleted:
+        if self.checkVar3.get():
+            ssplt.remove_last_image(self.canvas.figure.axes[0])
+        # If we activate it, print instructions
+        # Update the value
+        self.checkVar3.set(not self.checkVar3.get())
+        print('Draw mask :', self.checkVar3.get())
+        if self.checkVar3.get():
+            ax=self.canvas.figure.axes[0]
+            self.mask_plot = ax.imshow(self.mask, origin='lower', alpha=0.2, cmap='spring')
+
+        if self.checkVar3.get(): 
+            print('The mask has been activated. Now, move to the next frame and deactivate \
+                    it again so it does not add layers.')
         self.canvas.draw()