+++ ScintSuite/_StrikeMap/_FILD_INPA_ParentStrikeMap.py
@@ -14,6 +14,7 @@ import ScintSuite.errors as errors
 import ScintSuite._Plotting as ssplt
 from tqdm import tqdm
 from ScintSuite._Paths import Path
+from ScintSuite._SideFunctions import createGrid
 from ScintSuite.decorators import deprecated
 from ScintSuite._basicVariable import BasicVariable
 from ScintSuite._Mapping._Common import _fit_to_model_
@@ -726,6 +727,68 @@ class FILDINPA_Smap(GeneralStrikeMap):
                 # Update the header
                 strikes.header['info'].update(extra_column)
 
+    def  calculate_phaseSpace_to_pixelMatrix(self, gridPhaseSpace, gridPixel,
+                                             limitation: float = 10.0,
+                                             MC_number=300):
+        """
+        Calculate the transformation like matrix to go from the remap to the camera frame
+        
+        
+        """
+        # First calcualte the interpolators from phase space to pixel
+        # Select the variable to interpolate
+        xvar = self._data[gridPhaseSpace['xname']].data
+        print(xvar.mean())
+        yvar = self._data[gridPhaseSpace['yname']].data
+        print(yvar.mean())
+        gridPhaseSpace.pop('xname')
+        gridPhaseSpace.pop('yname')
+        
+        # Get the pixel position
+        xpix = self._coord_pix['x']
+        ypix = self._coord_pix['y']
+        
+        # Construct the phase space grid
+        nx, ny, xedges, yedges = createGrid(**gridPhaseSpace)
+        nxpix, nypix, xedgespix, yedgespix = createGrid(**gridPixel)
+        
+        # Construct the interpolators
+        interpolatorX = scipy_interp.LinearNDInterpolator(
+            np.column_stack((xvar.flatten(), yvar.flatten())),
+            xpix.flatten()
+        )
+        interpolatorY = scipy_interp.LinearNDInterpolator(
+            np.column_stack((xvar.flatten(), yvar.flatten())),
+            ypix.flatten()
+        )      # 
+        memory_size = nx * ny * nxpix * nypix \
+            * 8 / 1024 / 1024 / 1024
+        if memory_size > limitation:
+            text = 'The requiring matrix will consume %2.1f Gb, this is above'\
+                % memory_size\
+                + 'the threshold. Increase it if you really want to proceed'
+            raise errors.NotValidInput(text)
+        transform = np.zeros((nxpix, nypix, nx, ny), dtype='float64')
+        # Allocate the random number generator
+        rand = np.random.default_rng()
+        generator = rand.uniform
+        # Loop over the grid
+        for i in tqdm(range(nx)):
+            for j in range(ny):
+                # Create a set of MC markers
+                x_markers = generator(xedges[i], xedges[i+1], MC_number)
+                y_markers = generator(yedges[j], yedges[j+1], MC_number)
+                # Calculate the pixel position
+                xpix_markers= interpolatorX(x_markers, y_markers)
+                ypix_markers= interpolatorY(x_markers, y_markers)
+                # Histogram the pixel space
+                H, _, _ = np.histogram2d(xpix_markers, ypix_markers,
+                                         bins=[xedgespix, yedgespix])
+                # Save it in place
+                transform[:, :, i, j] = H.copy()
+        # Normalise the matrix
+        transform /= MC_number
+        return transform
     # --------------------------------------------------------------------------
     # --- Plotting Block
     # --------------------------------------------------------------------------
@@ -879,7 +942,7 @@ class FILDINPA_Smap(GeneralStrikeMap):
                                          kind_of_plot: str = 'normal',
                                          include_legend: bool = False,
                                          XI_index=None,
-                                         normalize: bool = False):
+                                         grid: bool = False):
         """
         Plot the fits done to calculate the resolution
 
@@ -890,7 +953,7 @@ class FILDINPA_Smap(GeneralStrikeMap):
         :param  gyr_index: index, or arrays of indeces, of gyroradius to plot
         :param  pitch_index: index, or arrays of indeces, of pitches to plot,
             this is outdated code, please use XI_index instead
-        :param  gyroradius: gyroradius value of array of then to plot. If
+        :param  gyroradius: gyroradius value or array of them to plot. If
         present, gyr_index will be ignored
         :param  pitch: idem to gyroradius bu for the pitch
         :param  kind_of_plot: kind of plot to make:
@@ -901,7 +964,6 @@ class FILDINPA_Smap(GeneralStrikeMap):
             - just_fit: Just a line plot as the fit
         :param  include_legend: flag to include a legend
         :param  XI_index: equivalent to pitch_index, but with the new criteria
-        :param  normalize: normalize the output
         """
         # --- Initialise plotting options and axis:
         default_labels = {
@@ -914,9 +976,12 @@ class FILDINPA_Smap(GeneralStrikeMap):
                 'ylabel': 'Counts [a.u.]'
             }
         }
-        ax_options = {
-            'grid': 'both',
-        }
+        if grid:
+            ax_options = {
+                'grid': 'both',
+            }
+        else:
+            ax_options = {}
         ax_options.update(default_labels[var.lower()])
         ax_options.update(ax_params)
         if ax is None:
@@ -929,7 +994,7 @@ class FILDINPA_Smap(GeneralStrikeMap):
         # --- Localise the values to plot
         if gyroradius is not None:
             # test if it is a number or an array of them
-            if isinstance(gyroradius, (list, np.ndarray)):
+            if isinstance(gyroradius, (np.ndarray)):
                 gyroradius = gyroradius
             else:
                 gyroradius = np.array([gyroradius])
@@ -937,7 +1002,7 @@ class FILDINPA_Smap(GeneralStrikeMap):
             for i in range(index_gyr.size):
                 index_gyr[i] = \
                     np.argmin(np.abs(self.MC_variables[1].data - gyroradius[i]))
-            logger.debug('Found gyroradius: %.2f' % self.MC_variables[1].data)
+            logger.debug('Found gyroradius: %.2f' % self.MC_variables[1].data[index_gyr])
         else:
             # test if it is a number or an array of them
             if gyr_index is not None:
@@ -984,9 +1049,9 @@ class FILDINPA_Smap(GeneralStrikeMap):
                     deltax = x.max() - x.min()
                     x_fine = np.linspace(x.min() - 0.1 * deltax,
                                          x.max() + 0.1 * deltax)
-                    name = 'rl: ' + str(round(self.MC_variables[1].data[ir], 1))\
+                    name = 'rL: ' + str(round(self.MC_variables[1].data[ir], 1))\
                         + ' $\\lambda$: ' + \
-                        str(round(self.MC_variables[1].data[ip], 1))
+                        str(round(self.MC_variables[0].data[ip], 1))
                     normalization = \
                         self._resolutions['norm_' + var.lower()][ip, ir]
                     y = self._resolutions['fits_' + var.lower()][ip, ir].eval(