+++ ScintSuite/_StrikeMap/_ParentStrikeMap.py
@@ -412,7 +412,7 @@ class GeneralStrikeMap(XYtoPixel):
                                     color=line_options['color'],
                                     rotation=rotation_for_y_label)
 
-        plt.draw()
+        #plt.draw()
         return ax
 
     def plot_real(self, ax=None, marker_params: dict = {},
@@ -632,6 +632,9 @@ class GeneralStrikeMap(XYtoPixel):
     @property
     def variables(self):
         return self._header['variables']['name']
+
+    def keys(self):
+        return self._data.keys()
 
     def __call__(self, name):
         return self._data[name].data