+++ CHANGELOG.md
@@ -1,9 +1,16 @@
+
 # 1.3.11 Implementation of xarray in forward modelling
 - Capable of creating synthetic siganls with WF xarray
 - Change in phase space remapped synthetic signals
 - Camera frames, with implemention of diferent kind of noises and filters
 - Capability to compute synthetic signals for double collimators
 - smap class fit parameters matrix wrap for easy optimization
+
+# 1.3.10 Minor fixes
+- Bug solved in substract noise, last frame was not taken into account (bug introduced in 1.2.1)
+- Plot mask option added to the GUI_frames
+- Included option to load both remap and video
+- Fixed the rotation of the frames when plotting for MU01
 
 # 1.3.9 D3D implementation
 ## General
@@ -38,7 +45,7 @@
 - Updated export fild, for the case of single smap used in remap
 
 # 1.3.5 Shares path, Update netCDF library for MU
-- Updated basic pathjs of AUG FILD and INPa to the shares
+- Updated basic paths of AUG FILD and INPa to the shares
 - The netCDF library has been updated to read new standard MU format
 
 # 1.3.4 Home flexibility